[501508] Remove assumption about model implementation class name
GraphicalEditPart.getAccessibleEditPart() assumes all the
implementation classes of the semantic model elements end with the
"Impl" suffix. This is not necessarily true if the EMF generators have
been customized, or if sub-classes with different names are used to
isolate non-generated code (for example Sirius uses XXXSpec classes
which override the generated XXXImpl with custom code).

When called on a class which does not match this assumption,
getAccessibleEditPart() crashes with a StringIndexOutOfBoundsException
because it fails to check the result of indexOf().

This code branch is only used as a fallback when a proper label is not
available, and the EClass name is certainly not worse than a raw Java
class name.

Bug: 501508
Change-Id: Iea77f77a64561a1c9dcf7f76a2e4da2c3c58968a
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>