[560031] Ignore jump links for overlapped segments
If the starting point, or ending point, is the same on the 2 lines, the
jump link is ignored because a part of the lines is common and the jump
link is not desirable as it does not improve the readability of the
diagram. This kind of jump links is called "partial jump links".

Bug: 560031
Change-Id: Idcbca62c936645f51886f32e2ebed90ec13d418a
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>