[566315] Improve clipboard image copy support
Image clipboard copy was only ever supported on Windows, but even on
Windows is broken on Java 11 or later.

By default, disable it when running on Java 11 or later, even on Windows.

A new system property 'org.eclipse.gmf.runtime.image.copy.supported'
can be used to explicitly enable or disable the feature if needed. It
overrides all the other tests.

Bug: 566315
Change-Id: I4e0e7a6bdb0e78fdef8b34bfde3bee555b505149
Signed-off-by: Axel RICHARD <axel.richard@obeo.fr>
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>