Fix compilation errors with the upcoming 2021-03 platform
In the upcoming version of the platform, IResource has a new method
createMarker() (see bug #570914). The interface is not supposed to be
implemented by clients, but the GMF Runtime tests have two "mock"
implementations (of IFile and IFolder) for testing purposes, and they
fail to compile when targeting the 'next' target platform.

See-also: https://bugs.eclipse.org/bugs/show_bug.cgi?id=570914
Change-Id: I18527298d250dcd648f1ff69b98ba77802de456b
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>