Hard-code the RGB values of 'black' in AppearancePreferencePage
We only access ColorConstants.black to get its RGB components, but
black is... black. In particular it does not depend on the current
theme.

Removing the reference to GEF's ColorConstants classes here may fix
https://github.com/eclipse/gef-classic/issues/201 and
https://bugs.eclipse.org/bugs/show_bug.cgi?id=581906 as it would avoid
initializing the whole ColorConstants class (including the
theme-dependent values) too early when
org.eclipse.gmf.runtime.diagram.ui is initliazed during an early
startup (as is the case in Papyrus).

Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>